requirements.txt
# ABOUTME: Core dependencies for MediaMTX control plane Flask application
# ABOUTME: Includes Flask, database, authentication, and security packages

# Flask core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1

# LDAP Authentication
ldap3==2.9.1

# Database
SQLAlchemy==2.0.23
# PostgreSQL driver (optional - only needed if using PostgreSQL)
# Uncomment the line below if using PostgreSQL:
# psycopg2-binary==2.9.9

# Security & Auth
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2
secrets-manager

# HTTP & API
requests==2.31.0
Werkzeug==3.0.1

# Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
